Refactor wallet encrypt/decrypt
